automation_hitss/requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
attrs==25.3.0
colorama==0.4.6
greenlet==3.2.3
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
playwright==1.52.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.1
pytest==8.4.0
pytest-html==4.1.1
pytest-metadata==3.1.1
typing_extensions==4.14.0
